Refactor FileTypeIconService to use filename instead of asset
The FileTypeIconService now expects a filename instead of a full-
blown Asset object for providing a matching file type icon.

The reason for this change is, that in some cases an Asset object is
not (yet) available, for example when dealing with Asset Proxy objects
from external asset sources.

The FileTypeIconService does not belong to the public API and therefore
the change isn't considered breaking backwards-compatibility.
