+++ Neos.Media.Browser/Resources/Private/Partials/ThumbnailView.html
@@ -35,7 +35,7 @@
                         </f:else>
                     </f:if>
                     <div class="neos-img-label">
-                        <mediaBrowser:fileTypeIcon filename="{assetProxy.filename}" height="32"/>
+                        <m:fileTypeIcon filename="{assetProxy.filename}" height="32"/>
                         <f:if condition="!{activeAssetSource.readOnly}">
                             <div class="neos-dropdown neos-pull-right" id="neos-asset-actionmenu-{assetProxy.identifier}">
                                 <a class="dropdown-toggle neos-button" href="#" data-neos-toggle="dropdown" data-target="#neos-asset-actionmenu-{assetProxy.identifier}">
